// frontend/src/api/generated/TripDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BookingDto } from "./BookingDto";

export type TripDto = { id: string, project_id: string, title: string, arrive_date: string, depart_date: string, participants: Array<string>, status: string, bookings: Array<BookingDto>, created_at: string, };